Added Resource based authorization

+++ Restaurants.Application/Dishes/Queries/DeleteDishForRestaurant/DeleteDishForRestaurantHandler.cs
@@ -1,13 +1,16 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Queries.DeleteDishForRestaurant;
 
-public class DeleteDishForRestaurantHandler(ILogger<DeleteDishForRestaurantHandler> logger, IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteDishForRestaurantCommand> 
+public class DeleteDishForRestaurantHandler(ILogger<DeleteDishForRestaurantHandler> logger,
+    IRestaurantAuthorizationService restaurantAuthorizationService, IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteDishForRestaurantCommand> 
 
 {
     public async Task Handle(DeleteDishForRestaurantCommand request, CancellationToken cancellationToken)
@@ -15,7 +18,10 @@ public class DeleteDishForRestaurantHandler(ILogger<DeleteDishForRestaurantHandl
         logger.LogInformation($"Deleting restaurant - {request.RestaurantId}");
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
         if (restaurant is null)
-            throw new NotFoundException(nameof(Restaurant), $"request.RestaurantId");
+            throw new NotFoundException(nameof(Restaurant), $"{request.RestaurantId}");
+        
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbiddenException(nameof(Restaurant), $"{request.RestaurantId}");
 
         await restaurantsRepository.DeleteAsync(restaurant);
     }
